Fix tests after #462
